--- packages/client/src/components/Buttons/Buttons.styles.css
@@ -1,139 +1,0 @@
-/* Buttons Style */
-.storybook-buttons {
-  font-family: 'Montserrat', Helvetica, Arial, sans-serif;
-  font-style: normal;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  text-align: center;
-  width: fit-content;
-}
-
-.storybook-button--submit {
-  width: 21.563rem;
-  height: 4.125rem;
-  background: #53742a;
-  border: 1px solid #ffffff;
-  color: #ffffff;
-  display: flex;
-  align-items: center;
-  font-weight: 500;
-  font-size: 1.875rem;
-}
-
-.storybook-button--small {
-  font-size: 0.75rem;
-  padding: 0.625rem 1rem;
-}
-
-.add-to-cart-svg {
-  width: 1.25rem;
-  height: 25rem;
-  padding-left: 0.625rem;
-}
-
-.storybook-button--submitSecondary {
-  width: 21.563rem;
-  height: 4.125rem;
-  background: #231c16;
-  border: 1px solid #ffffff;
-  color: #ffffff;
-  display: flex;
-  align-items: center;
-  font-weight: 500;
-  font-size: 1.875rem;
-}
-.storybook-button--shop {
-  width: 21.563rem;
-  height: 4.125rem;
-  background: #231c16;
-  border: 1px solid #ffffff;
-  color: #ffffff;
-  font-weight: 500;
-  font-size: 1.875rem;
-}
-
-.storybook-button--editDetails {
-  width: 12.058rem;
-  height: 2.313rem;
-  background-color: #2c251e;
-  border: 1px solid #ffffffd3;
-  color: #ffffff;
-  text-align: center;
-  font-weight: 400;
-  font-size: 0.938rem;
-}
-
-.storybook-button--continueShopping {
-  width: 13.313rem;
-  height: 2.313rem;
-  background: rgba(35, 28, 22, 0.9);
-  border: 1px solid #ffffff;
-  color: #ffffff;
-  font-weight: 400;
-  font-size: 0.938rem;
-  text-align: center;
-}
-
-.storybook-button--continueShoppingSecondary {
-  width: 12.058rem;
-  height: 2.313rem;
-  background: rgba(35, 28, 22, 0.9);
-  border: 1px solid #ffffff;
-  color: #ffffff;
-  font-weight: 400;
-  font-size: 0.938rem;
-  line-height: 1.125rem;
-  text-align: center;
-  padding: 0px;
-}
-
-.storybook-button--reviewOrderDetails {
-  width: 12.058;
-  height: 2.313rem;
-  background: rgba(35, 28, 22, 0.9);
-  border: 1px solid #ffffff;
-  color: #ffffff;
-  font-weight: 400;
-  font-size: 0.938rem;
-  line-height: 1.125rem;
-  text-align: center;
-}
-
-.storybook-button--checkout {
-  width: 13.313rem;
-  height: 37px;
-  background: #53742a;
-  border: 1px solid #ffffff;
-  color: #ffffff;
-  font-weight: 600;
-  font-size: 1.25rem;
-  line-height: 1.5rem;
-  text-align: center;
-}
-
-.storybook-button--reviewOrder {
-  width: 13.313rem;
-  height: 2.313rem;
-  background: #53742a;
-  border: 1px solid #ffffff;
-  color: #ffffff;
-  font-weight: 600;
-  font-size: 0.938rem;
-  line-height: 1.5rem;
-  text-align: center;
-}
-
-.storybook-button--addToCart {
-  width: 13.313rem;
-  height: 2.313rem;
-  background: #53742a;
-  border: 1px solid #ffffffd3;
-  color: #ffffff;
-  font-style: normal;
-  font-weight: 600;
-  font-size: 0.938rem;
-  line-height: 1.125rem;
-  text-align: center;
-  padding-left: 25px;
-}
